Update README.md
This Python script adeptly combines data manipulation with pandas, visualization with Matplotlib and Seaborn, statistical modeling with SciPy, and interactive plotting with Plotly to execute comprehensive financial analysis. The data is imported from a CSV file and thoroughly examined through exploratory analysis, which includes summary statistics and the handling of null values. Date conversions are
seamlessly performed, and the dataset is enriched with derived features such as the day of the week. Advanced preprocessing techniques like linear interpolation and mean imputation are employed to address missing values, while string representations of numbers are converted into floats for precise numerical analysis. The culmination of this process is a set of rich visualizations of stock prices and
volumes over time and the application of polynomial regression to model and forecast stock price trends, all structured into a pandas DataFrame for robust and insightful data processing.The use of the requests library to interact with the Alpha Vantage API, retrieving daily time-series data for notable stocks like Amazon (AMZN), Microsoft (MSFT), and Apple (AAPL).